% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{femData}
\alias{femData}
\title{Female Mice Liver Expression Data}
\format{
A data frame with 66 rows and 3600 columns
}
\usage{
data(femData)
}
\description{
This dataset includes gene expression levels from liver tissue
of female mice. The data originates from research conducted by
Ghazalpour et al (\url{https://doi.org/10.1186/1471-2105-9-559}).
It is also widely used as an WGCNA tutorial. However, at this moment,
the documentations for WGCNA no longer exist. See the original
WGCNA paper for more information
(\url{https://doi.org/10.1186/1471-2105-9-559}).
}
\keyword{datasets}
