% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{interaction}
\alias{interaction}
\title{Potential interaction variables for fuzzy_forest object}
\usage{
interaction(object, thresh)
}
\arguments{
\item{object}{A fuzzy_forest object.}

\item{thresh}{Threshold for interaction strength. Threshold should
be between \code{0} and \code{1} where the higher values
indicates strong interaction effects.}

\item{...}{Additional arguments not in use.}
}
\value{
A dataframe of potential pairwise interactions
}
\description{
Calculates potential interaction variables from the fuzzy_forest
object for a given threshold
}
