% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Liver_Exp}
\alias{Liver_Exp}
\title{Cleaned and Preprocess Female Mice Liver Expression Data}
\format{
A data frame with 66 rows and 3601 columns
}
\usage{
data(liverExp)
}
\description{
Removes outlier genes through WGCNA preprocessing of the Female
Liver Expression Data. Weight (g) variable from Clinical Trait
Data is also included. See Female Mice Liver Expression Vignette
for more information.
}
\keyword{datasets}
