% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapff_plot.R
\name{plot_interactions}
\alias{plot_interactions}
\title{Plot Interaction Plot from SHAP Fuzzy Forest}
\usage{
plot_interactions(object, ...)
}
\arguments{
\item{object}{A SHAP Fuzzy Forest object.}
}
\value{
A ggplot2 object representing the SHAP potential interaction matrix.
}
\description{
Generates an interaction plot from SHAP interaction values.
Interactions is generated from \code{shapviz}. Note this requires
setting \code{shap_type} = \code{tree} when running \code{shapff} or \code{shapwff}.
}
