% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapff_object.R
\name{predict.bonsai_forest}
\alias{predict.bonsai_forest}
\title{Predict from a Bonsai Forest Object}
\usage{
\method{predict}{bonsai_forest}(object, new_data, ...)
}
\arguments{
\item{object}{A \code{bonsai_forest} object.}

\item{new_data}{A matrix or data frame of new observations. Column names
must match the training features.}

\item{...}{Additional arguments passed to \code{ranger:::predict}.}
}
\value{
A \code{ranger} prediction object.
}
\description{
Generates predictions from the final random forest fitted on the stable
feature set.
}
