% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapff_object.R
\name{print.bonsai_forest}
\alias{print.bonsai_forest}
\title{Print a Bonsai Forest Object}
\usage{
\method{print}{bonsai_forest}(x, ...)
}
\arguments{
\item{x}{A \code{bonsai_forest} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Prints the stable feature set with mean |SHAP| importance scores and
shadow-stability frequencies.
}
