% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapff_object.R
\name{print.mossy_forest}
\alias{print.mossy_forest}
\title{Print a Mossy Forest Object}
\usage{
\method{print}{mossy_forest}(x, ...)
}
\arguments{
\item{x}{A \code{mossy_forest} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Prints the stable feature set with mean |SHAP| importance scores and
shadow-stability frequencies.
}
