% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapff_object.R
\name{print.shap_fuzzy_forest}
\alias{print.shap_fuzzy_forest}
\title{Prints output from fuzzy forests algorithm.}
\usage{
\method{print}{shap_fuzzy_forest}(x, ...)
}
\arguments{
\item{x}{A fuzzy_forest object.}

\item{...}{Additional arguments not in use.}
}
\value{
data.frame with list of selected features and variable
importance measures.
}
\description{
Prints output from fuzzy forests algorithm.
}
\note{
This work was partially funded by NSF IIS 1251151.
}
