% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shap_rfe.R
\name{select_RFE}
\alias{select_RFE}
\title{Selection step of shapley forest through SHAP values.}
\usage{
select_RFE(
  X,
  y,
  drop_fraction,
  shap_model = "full",
  number_selected,
  CLASSIFICATION,
  mtry_factor,
  ntree_factor,
  min_ntree,
  num_processors,
  nodesize,
  nsim,
  seed
)
}
\arguments{
\item{X}{A data.frame. With columns denoting a feature vector.
Could include additional covariates not a part of
the original modules.}

\item{y}{A Response vector.}

\item{drop_fraction}{A value between 0 and 1 representing the percentage of features
to be dropped at each iteration.}

\item{shap_model}{Binary indicator for \code{sf} model. If \code{full}, \code{sf}
runs SHAPley values at both screening and selection step.
If \code{after}, \code{sf} model runs SHAPley values at the end
of final model and keeps permutation VIMs usage at other steps.
\code{full} is default.}

\item{number_selected}{Number of features selected at the end of shapley forest.}

\item{CLASSIFICATION}{Binary indicator if the model is a classification model.}

\item{mtry_factor}{Number to adjust \code{mtry} for random forest.
If regression, \code{mtry} is set to
\code{ceiling}(\eqn{\sqrt(p)}*\code{mtry_factor}).
If classication, \code{mtry} is set to
\code{ceiling}((p/3)*\code{mtry_factor}).  If either
of these numbers is greater than p, \code{mtry} is
set to p.}

\item{ntree_factor}{A number greater than 1 to adjust \code{ntree}.  \code{ntree} for each
random is \code{ntree_factor} times the number
of features.  For each random forest, \code{ntree}
is set to \code{max}(\code{min_ntree},
\code{ntree_factor}*\code{p}).}

\item{min_ntree}{Minimum number of trees grown in each random forest.}

\item{num_processors}{Number of processors used to fit random forests.}

\item{nodesize}{Minimum \code{nodesize}}

\item{nsim}{Number of Monte Carlo repetitions for estimating SHAP
values in the screening step. Default is \code{1}. Increasing
\code{nsim} leads to more accurate results, but at the cost
of computational cost.}

\item{seed}{Seed to be set for reproducibility. Note: for a given seed,
parallel and non-parallel versions will differ.}
}
\value{
A data.frame with final surviving features.
}
\description{
Runs the selection step of shapley forest algorithm.
Returns data.frame with the final surviving SHAP values and features.
}
