% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapff_object.R
\name{shap_fuzzy_forest}
\alias{shap_fuzzy_forest}
\title{SHAP Fuzzy Forest Object}
\usage{
shap_fuzzy_forest(
  final_rf,
  final_X,
  module_membership,
  WGCNA_object = NULL,
  survivor_list,
  selection_list,
  final_shap,
  shap_obj
)
}
\arguments{
\item{final_rf}{A final random forest fit using the features
selected by fuzzy forests.}

\item{module_membership}{Module membership of each feature.}

\item{WGCNA_object}{If applicable, output of WGCNA analysis.}

\item{survivor_list}{List of features that have survived screening step.}

\item{selection_list}{List of features retained at each iteration of
selection step.}

\item{final_shap}{List of survived feature along with its SHAP
values}

\item{shap_obj}{Final shap object from the final surviving features}
}
\value{
An object of type fuzzy_forest.
}
\description{
Fuzzy forests with SHAPley values returns an object of type.
fuzzyforest.
}
