% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{traitData}
\alias{traitData}
\title{Female Mice Clinical Trait Data}
\format{
A data frame with 38 rows and 361 columns
}
\usage{
data(traitData)
}
\description{
This dataset goes a long with the Female Mice Liver Expression
Data. It contains clinical trait measurements of female mice.
It shares the same observations as Female Mice Liver Expression
Data and provides possible response variables for liver expression.
}
\keyword{datasets}
