% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapff_plot.R
\name{plot_decisions}
\alias{plot_decisions}
\alias{plot_decisions.shap_fuzzy_forest}
\title{Plot Decision Plot from SHAP Fuzzy Forest}
\usage{
plot_decisions(object, ...)

\method{plot_decisions}{shap_fuzzy_forest}(
  object,
  highlight = NULL,
  plot_title = "Decision Plot",
  geom_point = FALSE,
  gradient = c("blue", "red"),
  ...
)
}
\arguments{
\item{object}{A SHAP Fuzzy Forest object.}
}
\value{
A ggplot2 object representing the SHAP potential interaction matrix.
}
\description{
TO DO
}
\section{Methods (by class)}{
\itemize{
\item \code{plot_decisions(shap_fuzzy_forest)}: TO DO.

}}
