% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapff_plot.R
\name{plot_decisions}
\alias{plot_decisions}
\alias{plot_decisions.shapley_forest}
\title{Plot Decision Plot from shapley forest.}
\usage{
plot_decisions(object, ...)

\method{plot_decisions}{shapley_forest}(
  object,
  highlight = NULL,
  just_shap = FALSE,
  plot_title = "Decision Plot",
  geom_point = FALSE,
  gradient = c("blue", "red"),
  ...
)
}
\arguments{
\item{object}{A \code{shapley_forest} object. If \code{just_shap} is \code{TRUE}, a
\code{data.frame} of shapley values is expected instead.}

\item{...}{Obsolete additional arguments.}

\item{highlight}{A list of final surviving features to display for
decision plot. If \code{highlight} is \code{NULL}, all final
surviving features will be used}

\item{just_shap}{If \code{just_shap} is \code{TRUE}, the \code{object} is expected to only
contain a \code{data.frame} of shapley values--not a \code{shapley_forest}
object.}

\item{plot_title}{Optional. A name for decision plot. Default is \verb{Decision Plot}.}

\item{geom_point}{If \code{TRUE}, it will denote a point at each feature for each
observation. Default is \code{FALSE}.}

\item{gradient}{A vector denoting the starting color and ending color for
the gradient of the predicted outputs. Defaults is \code{blue}to
\code{red}.}
}
\value{
A ggplot2 object representing the decision plot.
}
\description{
Creates a decision plot from a \code{shapley_forest} object. Plot is adapted from
Python's \code{shap} package.
}
\section{Methods (by class)}{
\itemize{
\item \code{plot_decisions(shapley_forest)}: Decision plot for an object of class \code{shapley_forest} adapted from Python's \code{shap}.

}}
