% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapff_plot.R
\name{plot_dependence}
\alias{plot_dependence}
\alias{plot_dependence.shap_fuzzy_forest}
\title{Plot Dependence for Selected Features in SHAP Fuzzy Forests}
\usage{
plot_dependence(object, ...)

\method{plot_dependence}{shap_fuzzy_forest}(object, features, color_var = "auto", color = "#3b528b", ...)
}
\arguments{
\item{object}{A SHAP Fuzzy Forest object.}

\item{features}{The features for which the SHAP dependence plot should be generated.
Can be character or list.}

\item{color_var}{The variable used to color the points in the plot. Default is "auto", which
automatically selects a variable. For no color axis, set to \code{NULL}.
Can be character or list. See \code{shapviz} \link[shapviz]{sv_dependence()}
for more details.}

\item{color}{The color to use for the points in the dependence plot. Must be a valid
ggplot color or hex code. Default is "#3b528b".}
}
\value{
A \code{\link[ggplot2]{ggplot2}} object representing the dependence plot.
}
\description{
Generates a SHAP dependence plot for selected feature(s) in a SHAP Fuzzy Forest object.
Scatter plot of SHAP values against the selected feature(s). Plots are created
from \link[shapviz]{shapviz}.
}
\section{Methods (by class)}{
\itemize{
\item \code{plot_dependence(shap_fuzzy_forest)}: Dependence plot for an object of class "shap_fuzzy_forest" through \code{shapviz}.

}}
