% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapff_plot.R
\name{plot_elbow}
\alias{plot_elbow}
\alias{plot_elbow.shapley_forest}
\title{Plot Elbow Plot from shapley forest}
\usage{
plot_elbow(object, ...)

\method{plot_elbow}{shapley_forest}(
  object,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  colorName = NULL,
  color = TRUE,
  module_labels = NULL,
  ...
)
}
\arguments{
\item{object}{A \code{shapley_forest} object.}

\item{...}{Obsolete additional arguments.}

\item{main}{Main title for module plot. Default is \code{NULL} and will title
"Elbow Plot of Absolute SHAP values"}

\item{xlab}{Title for x-axis. Default is \code{NULL} and will be titled "Features".}

\item{ylab}{Title for y-axis. Default is \code{NULL} and will be titled
"Absolute SHAP Values"}

\item{colorName}{Title for module membership name. Only will display if \code{color} == \code{TRUE}.
Default is \code{NULL} and will be titled "Modules"}

\item{color}{Binary variable. Colors each SHAP value by the module color.
Default is \code{TRUE}.}

\item{module_labels}{Optional. A list of labels for the modules. Default is \code{NULL}.}
}
\value{
A ggplot object representing the module membership distribution colored by importance.
}
\description{
Generates a elbow plot of Absolute SHAP values of the final SHAP values.
}
\section{Methods (by class)}{
\itemize{
\item \code{plot_elbow(shapley_forest)}: Module plot for an object of class \code{shapley_forest} adapted from \code{fuzzyforest}.

}}
