% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapff_plot.R
\name{plot_force}
\alias{plot_force}
\alias{plot_force.shapley_forest}
\title{Plot SHAP Force Plot for Specific Observations in shapley forest}
\usage{
plot_force(object, ...)

\method{plot_force}{shapley_forest}(
  object,
  row_id,
  row_name = NULL,
  max_display = NULL,
  fill_colors = c("#59c46b", "#3b528b"),
  contrast = TRUE,
  bar_label_size = 3.2,
  show_annotation = TRUE,
  ...
)
}
\arguments{
\item{object}{A \code{shapley_forest} object.}

\item{...}{Obsolete additional arguments.}

\item{row_id}{The row index (or indices) corresponding to the observation(s) to
be generated on the waterfall plot. Can be character or list.
If list, selected observations are averaged.}

\item{row_name}{Optional. A name or vector of names for the observations.
If not provided, the \code{row_id} will be used.}

\item{max_display}{The maximum number of features to display in the plot.
If set to \code{NULL} it is defaulted to maximum possible.
Default is \code{NULL}.}

\item{fill_colors}{A vector of colors to fill the bars in the force plot.
Each color must be a valid ggplot color or hex code. Default is
\code{c("#59c46b","#3b528b")}.}

\item{contrast}{Whether to apply contrast between text and arrow. Default is \code{TRUE}.}

\item{bar_label_size}{The size of the labels on the bars. Default is \code{3.2}.}

\item{show_annotation}{Whether to show annotations for each bar in the waterfall plot.
Default is \code{TRUE}.}
}
\value{
A ggplot2 object representing the SHAP force plot.
}
\description{
Generates a force plot for specific observations in a \code{shapley_forest} object.
Plots are created from \link[shapviz]{shapviz}.
}
\section{Methods (by class)}{
\itemize{
\item \code{plot_force(shapley_forest)}: Force plot for an object of class \code{shapley_forest} through \code{shapviz}.

}}
