% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapff_plot.R
\name{plot_importance}
\alias{plot_importance}
\alias{plot_importance.shap_fuzzy_forest}
\title{Plot Feature Importance for SHAP Fuzzy Forests}
\usage{
plot_importance(object, ...)

\method{plot_importance}{shap_fuzzy_forest}(
  object,
  kind = "beeswarm",
  color_bar_title = "Feature Value",
  max_display = NULL,
  fill = "#3e568a",
  sort_features = TRUE,
  show_numbers = TRUE,
  viridis_args = list(begin = 0.25, end = 0.75, option = "viridis"),
  ...
)
}
\arguments{
\item{object}{A SHAP Fuzzy Forest object.}

\item{kind}{The type of plot to generate. Options are "bar", "beeswarm",
or "both". Default is "beeswarm".}

\item{color_bar_title}{Title for the color bar in the beeswarm plot. Default is "Feature Value".
Leave as is if \code{kind} = "bar".}

\item{max_display}{The maximum number of features to display.
If set to \code{NULL} it is defaulted to maximum possible.
Default is \code{NULL}.}

\item{fill}{The color to fill bars in the bar plot or color points in the beeswarm plot.
Must be a valid \code{\link[ggplot2]{ggplot2}} color or hex code.
Default is "#3e568a".}

\item{sort_features}{Whether to sort features by importance. Default is \code{TRUE}.}

\item{show_numbers}{Whether to show numeric SHAP values on the bar plot. Default is \code{TRUE}.}

\item{viridis_args}{Additional arguments passed to \code{viridis} for customizing the color scale
in the beeswarm plot.
Default is \code{list(begin = 0.25, end = 0.75, option = "viridis")}.}
}
\value{
A \code{\link[ggplot2]{ggplot2}} object representing the feature importance plot.
}
\description{
Generates a feature importance plot from SHAP Fuzzy Forest object. Three types
of plots can be generated: bar, beeswarm, and both plots. Plots are created
from \link[shapviz]{shapviz}.
}
\section{Methods (by class)}{
\itemize{
\item \code{plot_importance(shap_fuzzy_forest)}: Importance plot for an object of class "shap_fuzzy_forest" through \code{shapviz}.

}}
