% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapff_plot.R
\name{plot_potential_interactions}
\alias{plot_potential_interactions}
\alias{plot_potential_interactions.shapley_forest}
\title{Obtains Potential Interaction Matrix in shapley forest.}
\usage{
plot_potential_interactions(object, ...)

\method{plot_potential_interactions}{shapley_forest}(object, ...)
}
\arguments{
\item{object}{A \code{shapley_forest} object.}

\item{...}{Obsolete additional arguments.}
}
\value{
A ggplot2 object representing the SHAP potential interaction matrix.
}
\description{
Generates a potential interaction strength matrix of all features.
Interactions is calculated from \code{shapviz}'s
\link[shapviz]{potential_interactions}.
}
\section{Methods (by class)}{
\itemize{
\item \code{plot_potential_interactions(shapley_forest)}: Potential Interaction Matrix of class \code{shapley_forest} through \code{shapviz}.

}}
