% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapff_plot.R
\name{plot_potential_interactions}
\alias{plot_potential_interactions}
\alias{plot_potential_interactions.shap_fuzzy_forest}
\title{Plot Potential Interaction Matrix in SHAP Fuzzy Forests}
\usage{
plot_potential_interactions(object, ...)

\method{plot_potential_interactions}{shap_fuzzy_forest}(object, ...)
}
\arguments{
\item{object}{A SHAP Fuzzy Forest object.}
}
\value{
A ggplot2 object representing the SHAP potential interaction matrix.
}
\description{
Generates a potential interaction strength matrix of all features.
Interactions is calculated from \code{shapviz}'s
\link[shapviz]{potential_interactions}.
}
\section{Methods (by class)}{
\itemize{
\item \code{plot_potential_interactions(shap_fuzzy_forest)}: Potential Interaction Matrix of class "shap_fuzzy_forest" through \code{shapviz}.

}}
