% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapff_plot.R
\name{plot_waterfall}
\alias{plot_waterfall}
\alias{plot_waterfall.shap_fuzzy_forest}
\title{Plot Waterfall for Specific Observations in SHAP Fuzzy Forests}
\usage{
plot_waterfall(object, ...)

\method{plot_waterfall}{shap_fuzzy_forest}(
  object,
  row_id,
  row_name = NULL,
  max_display = NULL,
  order_fun = function(s) order(abs(s)),
  fill_colors = c("#59c46b", "#3b528b"),
  contrast = TRUE,
  show_connection = TRUE,
  show_annotation = TRUE,
  ...
)
}
\arguments{
\item{object}{A SHAP Fuzzy Forest object.}

\item{row_id}{The row index (or indices) corresponding to the observation(s) to
be generated on the waterfall plot. Can be character or list.
If list, selected observations are averaged.}

\item{row_name}{Optional. A name or vector of names for the observations.
If not provided, the \code{row_id} will be used.}

\item{max_display}{The maximum number of features to display in the plot.
If set to \code{NULL} it is defaulted to maximum possible.
Default is \code{NULL}.}

\item{order_fun}{Function used to order the features in the waterfall plot. Default orders
by absolute SHAP values.}

\item{fill_colors}{A vector of colors to fill the bars in the waterfall plot.
Each color must be a valid ggplot color or hex code. Default is
\code{c("#59c46b","#3b528b")}.}

\item{contrast}{Whether to apply contrast between text and arrow. Default is \code{TRUE}.}

\item{show_connection}{Whether to show connections between consecutive bars in the waterfall
plot. Default is \code{TRUE}.}

\item{show_annotation}{Whether to show annotations for each bar in the waterfall plot.
Default is \code{TRUE}.}
}
\value{
A \code{\link[ggplot2]{ggplot2}} object representing the waterfall plot.
}
\description{
Generates a waterfall plot for specific observation(s) in a SHAP Fuzzy Forest object.
Plots are  created from \link[shapviz]{shapviz}.
}
\section{Methods (by class)}{
\itemize{
\item \code{plot_waterfall(shap_fuzzy_forest)}: Waterfall plot for an object of class "shap_fuzzy_forest" through \code{shapviz}.

}}
