% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapff_object.R
\name{print.shapley_forest}
\alias{print.shapley_forest}
\title{Prints a summarized output of shapley forest.}
\usage{
\method{print}{shapley_forest}(x, ...)
}
\arguments{
\item{x}{\code{shapley_forest} object}

\item{...}{Obsolete additional arguments.}
}
\value{
data.frame with list of final selected features and its SHAP values.
}
\description{
Prints a summarized output of shapley forest.
}
