% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapff_object.R
\name{shapley_forest}
\alias{shapley_forest}
\title{SHAPley Forest Object}
\usage{
shapley_forest(
  final_rf,
  final_X,
  module_membership,
  WGCNA_object = NULL,
  survivor_list,
  selection_list,
  final_shap,
  shap_obj,
  runtime
)
}
\arguments{
\item{final_rf}{The final random forest object output.}

\item{final_X}{The final surviving \code{X}.}

\item{module_membership}{Module membership for each feature.}

\item{WGCNA_object}{WGCNA object output, if \code{wsf} was called.}

\item{survivor_list}{Feature list of surviving features after screening step.}

\item{selection_list}{Feature list of surviving features at each iteration
of the selection step.}

\item{final_shap}{Feature list of final SHAP features.}

\item{shap_obj}{Final shapley forest object from the final surviving features}

\item{runtime}{Data frame of runtimes of screening step, selection step,
and final random forest.}
}
\value{
An object of type \code{shapley_forest}.
}
\description{
Object return type for shapley forest algorithm.
}
