% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapselect_rf.R
\name{shapselect_RF}
\alias{shapselect_RF}
\title{The selection step of fuzzy forest algorithm}
\usage{
shapselect_RF(
  X,
  y,
  drop_fraction,
  number_selected,
  CLASSIFICATION,
  mtry_factor,
  ntree_factor,
  min_ntree,
  num_processors,
  nodesize,
  cl,
  nsim
)
}
\arguments{
\item{X}{A data.frame.
Each column corresponds to a feature vectors.
Could include additional covariates not a part of
the original modules.}

\item{y}{Response vector.}

\item{drop_fraction}{A number between 0 and 1.  Percentage of features
dropped at each iteration.}

\item{number_selected}{Number of features selected by fuzzyforest.}

\item{mtry_factor}{In the case of regression, \code{mtry} is set to
\code{ceiling}(\eqn{\sqrt(p)}*\code{mtry_factor}).
In the case of classification, \code{mtry} is set to
\code{ceiling}((p/3)*\code{mtry_factor}).  If either
of these numbers is greater than p, \code{mtry} is
set to p.}

\item{ntree_factor}{A number greater than 1.  \code{ntree} for each
random is \code{ntree_factor} times the number
of features.  For each random forest, \code{ntree}
is set to \code{max}(\code{min_ntree},
\code{ntree_factor}*\code{p}).}

\item{min_ntree}{Minimum number of trees grown in each random forest.}

\item{num_processors}{Number of processors used to fit random forests.}

\item{nodesize}{Minimum nodesize}

\item{cl}{Initialized cluster for parallelization. Set \code{cl}
to \code{NULL} if no parallelization is being ran.}

\item{nsim}{Number of Monte Carlo repetitions for estimating SHAP
values in the screening step. Default is \code{1}. Increasing
\code{nsim} leads to more accurate results, but at the cost
of computational cost.}
}
\value{
A data.frame with the top ranked features.
}
\description{
Carries out selection step of fuzzyforest algorithm using SHAP values.
Returns data.frame with variable importances and top rated features.
}
